Added Pagination

+++ app/controllers/comments_controller.rb
@@ -7,8 +7,10 @@ def create
   @comment.user_id = current_user.id
 
   if @comment.save
-    flash[:success] = "You commented the hell out of that post!"
-    redirect_back fallback_location: root_path
+    respond_to do |format|
+      format.html { redirect_to root_path }
+      format.js
+    end
   else
     flash[:alert] = "Check the comment form, something went horribly wrong."
     render root_path
@@ -19,13 +21,20 @@ def destroy
 
   @comment = @post.comments.find(params[:id])
 
-  if @comment.destroy
-    flash[:success] = "Comment Deleted!"
-    redirect_to root_path
-  else
-    flash[:alert] = "Post could not be deleted!"
+  if @comment.user_id == current_user.id
+    if @comment.destroy
+      respond_to do |format|
+        format.html { redirect_to root_path }
+        format.js
+      end
+    else
+      flash[:alert] = "Post could not be deleted!"
+    end
+  else 
+    flash[:alert] = "You can not delete other peoples comments!"
   end
 end
+
 private
 
 def comment_params
